+++ src/interceptors/transform.interceptor.ts
@@ -6,11 +6,11 @@
 
 import { Observable } from 'rxjs'
 import { map } from 'rxjs/operators'
-import { PaginateResult } from 'mongoose'
 import { Reflector } from '@nestjs/core'
 import { Injectable, NestInterceptor, CallHandler, ExecutionContext } from '@nestjs/common'
 import { HttpResponseSuccess, HttpPaginateResult, ResponseStatus } from '@app/interfaces/http.interface'
 import { ResponseMessage } from '@app/interfaces/http.interface'
+import { PaginateResult } from '@app/utils/paginate'
 import * as META from '@app/constants/meta.constant'
 import * as TEXT from '@app/constants/text.constant'
 
@@ -18,12 +18,12 @@ import * as TEXT from '@app/constants/text.constant'
 export function transformDataToPaginate<T>(data: PaginateResult<T>, request?: any): HttpPaginateResult<T[]> {
   return {
     data: data.docs,
-    params: request ? request.queryParams : null,
+    params: request?.queryParams || null,
     pagination: {
       total: data.total,
       current_page: data.page,
-      total_page: data.pages,
-      per_page: data.limit,
+      per_page: data.perPage,
+      total_page: data.totalPage,
     },
   }
 }
@@ -40,13 +40,15 @@ export class TransformInterceptor<T> implements NestInterceptor<T, HttpResponseS
     const call$ = next.handle()
     const target = context.getHandler()
     const request = context.switchToHttp().getRequest()
-    const message =
-      this.reflector.get<ResponseMessage>(META.HTTP_SUCCESS_MESSAGE, target) || TEXT.HTTP_DEFAULT_SUCCESS_TEXT
+    const message = this.reflector.get<ResponseMessage>(META.HTTP_SUCCESS_MESSAGE, target)
     const usePaginate = this.reflector.get<boolean>(META.HTTP_RES_TRANSFORM_PAGINATE, target)
     return call$.pipe(
       map((data: any) => {
-        const result = !usePaginate ? data : transformDataToPaginate<T>(data, request)
-        return { status: ResponseStatus.Success, message, result }
+        return {
+          result: usePaginate ? transformDataToPaginate<T>(data, request) : data,
+          status: ResponseStatus.Success,
+          message: message || TEXT.HTTP_DEFAULT_SUCCESS_TEXT,
+        }
       })
     )
   }