+++ src/interfaces/mongoose.interface.ts
@@ -4,7 +4,8 @@
  * @author Surmon <https://github.com/surmon-china>
  */
 
+import { Document } from 'mongoose'
 import { ModelType } from '@typegoose/typegoose/lib/types'
-import { PaginateModel, Document } from 'mongoose'
+import { PaginateModel } from '@app/utils/paginate'
 
 export type MongooseModel<T> = ModelType<T> & PaginateModel<T & Document>